Move 'SiddhiAppRuntimeService' into 'org.wso2.carbon.stream.processor.common' Remove unwanted dependency versions Add Editor Siddhi App Runtime Service
