+++ components/org.wso2.carbon.siddhi.store.api.rest/pom.xml
@@ -105,7 +105,7 @@
 
         <dependency>
             <groupId>org.wso2.carbon.analytics</groupId>
-            <artifactId>org.wso2.carbon.stream.processor.core</artifactId>
+            <artifactId>org.wso2.carbon.stream.processor.common</artifactId>
         </dependency>
 
         <dependency>
@@ -196,6 +196,7 @@
             org.wso2.msf4j.*;version="${msf4j.import.version.range}",
             org.wso2.siddhi.*;version="${siddhi.version.range}",
             org.wso2.transport.http.netty.config.*;version="${carbon.transport.package.import.version.range}",
+            org.wso2.carbon.stream.processor.common.*;version="${carbon.analytics.version.range}",
             org.wso2.carbon.analytics.msf4j.interceptor.common.*;version="${carbon.analytics.version.range}",
             *;resolution:=optional
         </import.package>
