+++ features/org.wso2.carbon.siddhi.store.api.rest.feature/pom.xml
@@ -59,10 +59,6 @@
             <artifactId>slf4j</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.wso2.carbon.analytics</groupId>
-            <artifactId>org.wso2.carbon.stream.processor.core</artifactId>
-        </dependency>
-        <dependency>
             <groupId>org.wso2.siddhi</groupId>
             <artifactId>siddhi-core</artifactId>
         </dependency>
